Don't show delete on books

+++ app/views/books/tab.rb
@@ -14,11 +14,13 @@ module Views
             div class: "flex items-center justify-end gap-2 py-2 px-4 bg-gray-200" do
               input type: "submit", value: "Update", form: "searchForm", formaction: view_path(@view.id), class: "inline-flex items-center rounded-md border border-transparent bg-blue-500 hover:bg-blue-400 text-white px-2.5 py-1.5 text-base font-medium text-gray-900 gap-2"
 
-              a "Delete", href: view_path(@view.id), data: {
-                turbo: {
-                  method: "delete"
+              unless @view.name == "Books"
+                a "Delete", href: view_path(@view.id), data: {
+                  turbo: {
+                    method: "delete"
+                  }
                 }
-              }
+              end
             end
           end
         end
